Adding more user stories
Adding a user story where organizers can see the attendance rate of the participants. There's a time where the seat is limited but got lots of people in whitelist. At the event day, some people doesn't come without a proper notice. This is a loss as there are some people that way more eager to join but probably got late notice of the event