Spec country code as 2-letter, use i18n-friendly column names for state/zip
The idea here would be to store countries as ISO 3166-1 alpha-2 strings for consistency. Also, the US is the only country with a ZIP code to my knowledge; everyone else calls them postal codes or postcodes. "State" is a little more widely used, but "region" is more correct for more places.