Mock search bar with Material UI. refs #349
This will require a call to `makeStyles()` in the future.
Also, the logo is too large when compared to the Prototype.
The spacing could be improved, if the placeholder text would be
shorter.
I'd suggest to apply a Floating Label pattern in the future for
accessibility and UX reasons:
https://csslayout.io/patterns/floating-label

My React + Material UI knowledge isn't strong enough yet to do so
on my own.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>