feat: compile ts incrementally (#685)
* chore: move db inside server

* chore: move typeorm inside server

* chore: move server deps into its package.json

* chore: fix ts-node/ts-node-dev issues

* chore: move tsconfig into server

* fix: update docker config

* fix: update postInstall to install the server

* test: run tsc in server dir

* test: stop running tests inside docker

Unit tests inside the containers don't tell us much and don't work now
that the test packages are not included in the app container

* test: run cypress tests against docker container

* chore: move config to src

* chore: move controllers to src

* chore: move services to src

* chore: move models to src

* chore: move util to src

* chore: move root files into src

* chore: move testUtils into test

* refactor: move gql types inside src

* fix: update import for production.ts

* feat: incremental compilation in dev mode

* chore: remove unused deps

* fix: compile on install

* chore: stop linting of generated files

* chore: move server dev-deps

* chore(deps): add concurrently to server

Since the root dependencies aren't getting into the docker image anymore
this adds concurrently to the server

* fix: make .env available to docker

Also makes the server's volume mirror the clients (i.e. it's just the
server dir, not root)

* feat: use app.js in production mode

production.js couples the client and server too tightly - it directly
requires the client's .next directory. This coupling makes it hard to
deploy the client and server to different VMs/containers.

* test: improve CI speed with DOCKER_BUILDKIT

* feat: create dev and ci docker configs

* test: use ci specific config in workflow

* fix: pass DOCKER_BUILDKIT env var correctly

* fix: wait for /graphql, since / 404s

* fix: cleanup

* fix: move and update dockerignores

* refactor: reduce script duplication

typeorm needs to be called from the server directory, but this can be
achieved by cd'ing into server rather than using npm scripts that call
npm scripts

* docs: add typeorm change + new models location

* docs: remove Docker specific test commands

Running the tests inside Docker seems redundant - if we make a mistake
in the Docker config, the e2e tests will flag that up.

* docs: npm i, not npx i