feat: use 'lax' cookies to get session in callback
'strict' cookies are not sent (well, should not be) when a browser
navigates to a domain rather than within a domain.  This means the user
is effectively logged out until the reload the page.

To work around this, allowing the code from Google to reach the server,
I've set the cookies to 'lax'. This should be okay, since the random
state should prevent CSRF