fix: allow venues to be changed to online
Having event's fk for venue be composite meant that the query to remove
that reference would try and null both keys (chapter_id and venue_id),
but chapter_id is not nullable.

This was a bit of a hack to require chapterId to be sent as a
GraphQL variable (so that authChecker could see it), this just opts for
the lesser of two hacks. Namely, passing, but not using, chapterId.