feat: allow calendar and event creation (#1442)
* feat: add keys dir

* feat: add Google oauth2 flow

* feat: use 'lax' cookies to get session in callback

'strict' cookies are not sent (well, should not be) when a browser
navigates to a domain rather than within a domain.  This means the user
is effectively logged out until the reload the page.

To work around this, allowing the code from Google to reach the server,
I've set the cookies to 'lax'. This should be okay, since the random
state should prevent CSRF

* refactor: cleanup

* refactor: use google-auth-library

* refactor: encapsulate google apis

* fix: use req.user for isLoggedIn

* feat: store tokens in the DB

* fix: use one OAuth2Client per request

Otherwise the wrong user's credentials might end up being used

* feat: allow user to set OAuth2 credentials

* chore(deps): downgrade google-auth-library

@googleapis/calendar relies on google-auth-library@7

* feat: add a createCalendar service

Also, moves OAuth2Client creation inside createCalendar. This gives
better encapsulation and ensures the correct token is used for the
request

* feat: create calendar on chapter creation

* feat: create service to create calendar events

* feat: create calendar event when creating an event

* feat: granting and revocation of calendar access

* refactor: pass variables differently

* feat: store calendar event id on creation

* fix: make missing keys only throw in production

* fix: suppress google service errors

* feat: update calendar when event is updated

* feat: cancel calendar event on event cancel

* feat: delete calendar event on event deletion

* feat: adds a link to authorize with Google

Also changes the POST endpoint to GET, so that the user can simply click
on a link and get redirected to Google's authorization server

* refactor: DRY out requestBody generation

* fix: stop guests seeing others or changing event

* feat: update event attendees on RSVP

* feat: only store one user's tokens

The calendars will all be owned by the first owner to authenticate with
Google.  Everyone else will use the owner's tokens to manipulate them.

* feat: store google account's email address

In case something goes horribly wrong.

* feat: reject authentication with invalid email

Ensures that the Google tokens never get overwritten by the wrong user's
tokens.

* chore: remove old comment

* feat: restrict auth routes to owners

* feat: hide Authenticate button from non-owners

* fix: keep attendee status (yes/no/maybe) on RSVP

Including start/end makes Google remove the user's status, but since
they're not different, it also doesn't email them with an update.

Using patch avoids this (and is neater)

* refactor: add comment about patch

* feat: remove calendar invite when RSVP is canceled

* feat: remove invite when RSVP deleted

* feat: invite user to event after confirming rsvp

* fix: use the updated event_users to get email list

* refactor: warn about temporal dependency

* fix: remove calendar access control and old code

* refactor: mistake in error message

Co-authored-by: Krzysztof G. <60067306+gikf@users.noreply.github.com>

* refactor: combine auth with api creation

 Co-authored-by: Krzysztof G. <60067306+gikf@users.noreply.github.com>

* refactor: move event creation into resolver

Co-authored-by: Krzysztof G. <60067306+gikf@users.noreply.github.com>

* refactor: rename updated event variable

Co-authored-by: Krzysztof G. <60067306+gikf@users.noreply.github.com>