fix: find events/venues with user.id in auth
Remarkably this was not a security issue.  The wrong sets of events
or venues would, at worst, cause the api to reject a request.

It's still bad, obviously, since users logging in and out repeated will
push up the session id and prevent users from making any requests
that require authentication.