fix: only require single logout request on Chrome
Some browsers (such as Brave) delete third party cookies after when the
page unloads (i.e. refresh or tab closing). Chrome, however, does not
and this means that isAuthenticated is true when useSession checks to
see if it should create or destroy the session.

By specifying the intent, we can ensure the session is deleted on logout