fix: only require chrome users to logout once (#2263)
* feat: stop trying to regenerate session

It's only useful in the weird edge case where a user has deleted their
session cookie, but not their auth0 cookies. In this case, they can just
login again.

* fix: only require single logout request on Chrome

Some browsers (such as Brave) delete third party cookies after when the
page unloads (i.e. refresh or tab closing). Chrome, however, does not
and this means that isAuthenticated is true when useSession checks to
see if it should create or destroy the session.

By specifying the intent, we can ensure the session is deleted on logout